* release-candidate: (742 commits)
  reap call sites before reaping the warm dashboard
  refines -F argument handling
  alphabetizes u3_opts and usage instructions
  replaces -N (enable fake networking) with -L (disable networking)
  removes the option to set -l, -n, and -r (raft options)
  removes -X ("skip last event")
  removes -M ("memory madness")
  Use Ed25519 u3a_malloc in the ed25519 jet.
  Fix README build instructions
  properly zero-initialize u3_wcon linked-list pointer
  removes second staging copy of compiler jets
  [CI] Point to latest release-candidate arvo
  De-duplicate dependency list
  Make small stylistic improvements to crypto jets
  meson v0.44.1
  Use fetchGit instead of fetchTarball to get nixpkgs.
  Revert change to libh2o for now -- Breaks the nix build, but unbreaks the normal build.
  Minor Cleanup
  Make list indentation consistent in Meson file.
  Use latest urbit/argon2
  ...

+++ vere/unix.c
@@ -1108,7 +1108,7 @@ u3_unix_acquire(c3_c* pax_c)
       kill(getpid(), SIGTERM);
       sleep(1); c3_assert(0);
     }
-    else {
+    else if (pid_w != getpid()) {
       c3_w i_w;
 
       if ( -1 != kill(pid_w, SIGTERM) ) {
